Organize tests per each page in different directories
